+++ python/GafferImageUI/OpenColorIOTransformUI.py
@@ -111,15 +111,10 @@ Gaffer.Metadata.registerNode(
 			> node instead.
 			""",
 
-			# We don't use the default CompoundDataPlugValueWidget because
-			# it allows the addition of all sorts of member plugs, and we
-			# only want to add strings. Instead we use the _ContextFooter
-			# to provide a button for only adding strings.
-			## \todo Perhaps we should invent some metadata scheme to give
-			# this behaviour to the CompoundDataPlugValueWidget?
 			"plugValueWidget:type" : "GafferUI.LayoutPlugValueWidget",
-			"layout:customWidget:addButton:widgetType" : "GafferImageUI.OpenColorIOTransformUI._ContextFooter",
+			"layout:customWidget:addButton:widgetType" : "GafferUI.PlugCreationWidget",
 			"layout:customWidget:addButton:index" : -1,
+			"plugCreationWidget:includedTypes" : "Gaffer.StringPlug",
 
 			"layout:section" : "Context",
 			"layout:index" : -3,
@@ -132,34 +127,3 @@ Gaffer.Metadata.registerNode(
 	}
 
 )
-
-class _ContextFooter( GafferUI.Widget ) :
-
-	def __init__( self, plug, **kw ) :
-
-		row = GafferUI.ListContainer( GafferUI.ListContainer.Orientation.Horizontal )
-		GafferUI.Widget.__init__( self, row, **kw )
-
-		with row :
-
-			GafferUI.Spacer( imath.V2i( GafferUI.PlugWidget.labelWidth(), 1 ) )
-
-			button = GafferUI.Button(
-				image = "plus.png",
-				hasFrame = False,
-				toolTip = "Click to add variables",
-			)
-
-			button.clickedSignal().connect( Gaffer.WeakMethod( self.__clicked ) )
-
-			GafferUI.Spacer( imath.V2i( 1 ), imath.V2i( 999999, 1 ), parenting = { "expand" : True } )
-
-		self.__plug = plug
-
-	def __clicked( self, button ) :
-
-		if Gaffer.MetadataAlgo.readOnly( self.__plug ) :
-			return
-
-		with Gaffer.UndoScope( self.__plug.ancestor( Gaffer.ScriptNode ) ) :
-			self.__plug.addChild( Gaffer.NameValuePlug( "", "", True, "member1", flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic ) )