+++ python/GafferSceneUI/CustomOptionsUI.py
@@ -34,9 +34,14 @@
 #
 ##########################################################################
 
+import functools
+
+import IECore
+
 import Gaffer
 import GafferUI
 import GafferScene
+import GafferSceneUI
 
 Gaffer.Metadata.registerNode(
 
@@ -63,7 +68,9 @@ Gaffer.Metadata.registerNode(
 			python.
 			""",
 
+			"plugCreationWidget:excludedTypes" : "Gaffer.ObjectPlug",
 			"compoundDataPlugValueWidget:editable" : True,
+			"ui:scene:acceptsOptions" : True,
 
 		},
 
@@ -96,3 +103,100 @@ Gaffer.Metadata.registerNode(
 	}
 
 )
+
+##########################################################################
+# PlugCreationWidget menu extension
+##########################################################################
+
+def __addFromGlobalsMenuDefinition( menu ) :
+
+	plugCreationWidget = menu.ancestor( GafferUI.PlugCreationWidget )
+	node = plugCreationWidget.plugParent().node()
+
+	result = IECore.MenuDefinition()
+
+	scene = next( GafferScene.ScenePlug.RecursiveInputRange( node ) )
+	options = scene["globals"].getValue()
+	existingNames = { plug["name"].getValue() for plug in plugCreationWidget.plugParent() }
+
+	prefix = "option:"
+
+	for name, value in [ ( k[len(prefix):], v ) for k, v in options.items() if k.startswith( prefix ) ] :
+		result.append(
+			"/{}".format( name ),
+			{
+				"command" : functools.partial( __createPlug, name = name, value = value ),
+				"active" : name not in existingNames
+			}
+		)
+
+	if not len( result.items() ) :
+		result.append(
+			"/No Options Found", { "active" : False }
+		)
+		return result
+
+	return result
+
+def __createPlug( menu, name, value ) :
+
+	plugCreationWidget = menu.ancestor( GafferUI.PlugCreationWidget )
+	plugCreationWidget.createPlug(
+		Gaffer.PlugAlgo.createPlugFromData( "plug0", Gaffer.Plug.Direction.In, Gaffer.Plug.Flags.Default, value ),
+		name = name
+	)
+
+def __plugCreationMenu( menuDefinition, widget ) :
+
+	if not Gaffer.Metadata.value( widget.plugParent(), "ui:scene:acceptsOptions" ) :
+		return
+
+	menuDefinition.prepend( "/FromSceneDivider", { "divider" : True } )
+
+	menuDefinition.prepend(
+		"/From Scene",
+		{
+			"subMenu" : __addFromGlobalsMenuDefinition
+		}
+	)
+
+GafferUI.PlugCreationWidget.plugCreationMenuSignal().connect( __plugCreationMenu )
+
+##########################################################################
+# PlugCreationWidget drag & drop extension
+##########################################################################
+
+def __filteredOptions( widget, dragDropEvent ) :
+
+	options = GafferSceneUI.SceneInspector.draggedOptions( dragDropEvent )
+	if not options :
+		return None
+
+	existingNames = { plug["name"].getValue() for plug in widget.plugParent() }
+	return {
+		k : v for k, v in options.items()
+		if k not in existingNames
+	}
+
+def __optionsDropHandler( widget, dragDropEvent ) :
+
+	options = __filteredOptions( widget, dragDropEvent )
+	if not options :
+		GafferUI.PopupWindow.showWarning( "Options added already", parent = widget )
+
+	with Gaffer.UndoScope( widget.plugParent().ancestor( Gaffer.ScriptNode ) ) :
+		for name, value in options.items() :
+			plug = Gaffer.PlugAlgo.createPlugFromData( "value", Gaffer.Plug.Direction.In, Gaffer.Plug.Flags.Default, value )
+			widget.createPlug( plug, name = name )
+
+def __plugCreationDragEnter( widget, dragDropEvent ) :
+
+	if not Gaffer.Metadata.value( widget.plugParent(), "ui:scene:acceptsOptions" ) :
+		return
+
+	if __filteredOptions( widget, dragDropEvent ) is not None :
+		return __optionsDropHandler
+
+	return None
+
+GafferUI.PlugCreationWidget.plugCreationDragEnterSignal().connect( __plugCreationDragEnter )