+++ Changes.md
@@ -6,6 +6,14 @@ Improvements
 
 - Timeline : Added scrubbing outside the playback range, enabled by holding <kbd>Shift</kbd>.
 - RenderMan : Added depth of field support (#6713).
+- NodeEditor :
+  - Added BoxPlug items to the plug creation menus.
+  - Added more Array items to the plug creation menus.
+  - Added drag & drop creation of plugs by dropping a plug or value onto the `+` icon.
+- AttributeTweaks : Added "Attributes" submenu to `name` plugs, matching the menu from CustomAttributes.
+- CustomAttributes : Added "From Selected" and "From Affected" submenus to the plug creation menu.
+- AttributeTweaks, OptionTweaks, CustomAttributes, CustomOptions, OptionQuery : Added drag & drop population of plugs by dragging from the Scene Inspector.
+- CustomOptions : Added "From Scene" submenu to the plug creation menu.
 
 Fixes
 -----
@@ -27,6 +35,9 @@ API
 - MenuButton : Added `set/getImmediate()` methods. These allow the button to execute a menu item immediately instead of showing the menu, if the menu has only a single item.
 - Signals : Added CatchingCombiner Python class, equivalent to the C++ template class.
 - FileSystemPath : Implemented Python's `os.PathLike` interface, to allow FileSystemPath objects to be passed directly to Python's standard filesystem functions.
+- PlugCreationWidget : Added standard metadata-driven widget to allow users to add plugs to nodes.
+- UserPlugs : Deprecated. Use PlugCreationWidget instead.
+- CompoundDataPlugValueWidget : Deprecated. Use LayoutPlugValueWidget and PlugCreationWidget instead.
 
 Build
 -----